refactor: rename "common/proto" to "common/commonpb"
No import aliases.
