+++ common/meson.build
@@ -4,4 +4,4 @@ lib_common_dep = declare_dependency(
   include_directories : [yanet_rootdir,],
 )
 
-subdir('proto')
+subdir('commonpb')
