+++ controlplane/ynpb/function.proto
@@ -4,7 +4,7 @@ package ynpb;
 
 option go_package = "github.com/yanet-platform/yanet2/controlplane/ynpb;ynpb";
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 // TODO: docs.
 service FunctionService {
