+++ devices/vlan/cli/build.rs
@@ -1,7 +1,7 @@
 use core::error::Error;
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../../../common/proto/target.proto");
+    println!("cargo:rerun-if-changed=../../../common/commonpb/target.proto");
     println!("cargo:rerun-if-changed=../controlplane/vlanpb/vlan.proto");
 
     tonic_build::configure()
@@ -9,7 +9,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
         .build_server(false)
         .message_attribute(".", "#[derive(Serialize)]")
         .compile_protos(
-            &["common/proto/target.proto", "vlanpb/vlan.proto"],
+            &["common/commonpb/target.proto", "vlanpb/vlan.proto"],
             &["../../..", "../controlplane"],
         )?;
