+++ devices/vlan/controlplane/vlanpb/vlan.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package vlanpb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/devices/vlan/controlplane/vlanpb;vlanpb";
