+++ modules/acl/controlplane/aclpb/acl.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package aclpb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/modules/"
 		    "acl/controlplane/aclpb;aclpb";
