+++ modules/balancer/controlplane/balancerpb/balancer.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package balancerpb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 import "google/protobuf/timestamp.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/modules/"
