+++ modules/balancer/controlplane/service.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 	"time"
 
-	commonpb "github.com/yanet-platform/yanet2/common/proto"
+	"github.com/yanet-platform/yanet2/common/commonpb"
 	"github.com/yanet-platform/yanet2/controlplane/ffi"
 	"github.com/yanet-platform/yanet2/modules/balancer/controlplane/balancerpb"
 	"go.uber.org/zap"
