+++ modules/dscp/controlplane/dscppb/dscp.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package dscppb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/controlplane/modules/dscp/dscppb;dscppb";
