+++ modules/forward/controlplane/forwardpb/forward.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package forwardpb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/modules/forward/controlplane/forwardpb;forwardpb";
