+++ modules/nat64/cli/build.rs
@@ -1,7 +1,7 @@
 use core::error::Error;
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../../../common/proto/target.proto");
+    println!("cargo:rerun-if-changed=../../../common/commonpb/target.proto");
     println!("cargo:rerun-if-changed=../controlplane/nat64pb/nat64.proto");
 
     tonic_build::configure()
@@ -9,7 +9,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
         .build_server(false)
         .message_attribute(".", "#[derive(Serialize)]")
         .compile_protos(
-            &["common/proto/target.proto", "nat64pb/nat64.proto"],
+            &["common/commonpb/target.proto", "nat64pb/nat64.proto"],
             &["../../..", "../controlplane"],
         )?;
