+++ modules/nat64/controlplane/nat64pb/nat64.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package nat64pb;
 
-import "common/proto/target.proto";
+import "common/commonpb/target.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/modules/nat64/controlplane/nat64pb;nat64pb";
