+++ modules/pdump/cli/build.rs
@@ -2,7 +2,7 @@ use core::error::Error;
 use std::{env, path::PathBuf};
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../../../common/proto/target.proto");
+    println!("cargo:rerun-if-changed=../../../common/commonpb/target.proto");
     println!("cargo:rerun-if-changed=../controlplane/pdumppb/pdump.proto");
 
     tonic_build::configure()
@@ -10,7 +10,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
         .build_server(false)
         .message_attribute(".", "#[derive(Serialize)]")
         .compile_protos(
-            &["common/proto/target.proto", "pdumppb/pdump.proto"],
+            &["common/commonpb/target.proto", "pdumppb/pdump.proto"],
             &["../../..", "../controlplane"],
         )?;
